Added 'create room' functionality. Also added 'modify patient' and 'modify staff' functionalities. Finally also added success messages, fixed database and fixed minor errors in old code.

+++ rempat.py
@@ -28,12 +28,22 @@ def delpat():
         tmp = sp.call('clear', shell = True)
         pid = int(input("Enter Patient Id of patient to be removed: "))
 
+        query = """select * from Patient where Patient_Id = %d""" % (pid)
+        if(qexec(query)): 
+            return -1; 
+        res = cur.fetchall()
+        if(res == ()):
+            print("Invalid id. No such patient.")
+            return -1
+
         if(delins(pid)):
             return -1
 
         query = """delete from Patient where Patient_Id = %d""" % (pid)
         if(qexec(query)):
             return -1;
+        
+        print("Success!")
         return 0
 
     except Exception as e:
